+++ pkg/disttask/framework/storage/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
     deps = [
         "//pkg/disttask/framework/proto",
         "//pkg/kv",
-        "//pkg/parser/terror",
         "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/util/chunk",
