+++ pkg/disttask/framework/storage/task_table.go
@@ -28,7 +28,6 @@ import (
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/pkg/disttask/framework/proto"
 	"github.com/pingcap/tidb/pkg/kv"
-	"github.com/pingcap/tidb/pkg/parser/terror"
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
 	"github.com/pingcap/tidb/pkg/util/chunk"
@@ -113,20 +112,6 @@ func SetTaskManager(is *TaskManager) {
 	taskManagerInstance.Store(is)
 }
 
-// ExecSQL executes the sql and returns the result.
-// TODO: consider retry.
-func ExecSQL(ctx context.Context, se sessionctx.Context, sql string, args ...interface{}) ([]chunk.Row, error) {
-	rs, err := se.(sqlexec.SQLExecutor).ExecuteInternal(ctx, sql, args...)
-	if err != nil {
-		return nil, err
-	}
-	if rs != nil {
-		defer terror.Call(rs.Close)
-		return sqlexec.DrainRecordSet(ctx, rs, 1024)
-	}
-	return nil, nil
-}
-
 func row2TaskBasic(r chunk.Row) *proto.Task {
 	task := &proto.Task{
 		ID:          r.GetInt64(0),
@@ -183,7 +168,7 @@ func (stm *TaskManager) WithNewSession(fn func(se sessionctx.Context) error) err
 func (stm *TaskManager) WithNewTxn(ctx context.Context, fn func(se sessionctx.Context) error) error {
 	ctx = util.WithInternalSourceType(ctx, kv.InternalDistTask)
 	return stm.WithNewSession(func(se sessionctx.Context) (err error) {
-		_, err = ExecSQL(ctx, se, "begin")
+		_, err = sqlexec.ExecSQL(ctx, se, "begin")
 		if err != nil {
 			return err
 		}
@@ -194,7 +179,7 @@ func (stm *TaskManager) WithNewTxn(ctx context.Context, fn func(se sessionctx.Co
 			if success {
 				sql = "commit"
 			}
-			_, commitErr := ExecSQL(ctx, se, sql)
+			_, commitErr := sqlexec.ExecSQL(ctx, se, sql)
 			if err == nil && commitErr != nil {
 				err = commitErr
 			}
@@ -211,7 +196,7 @@ func (stm *TaskManager) WithNewTxn(ctx context.Context, fn func(se sessionctx.Co
 
 func (stm *TaskManager) executeSQLWithNewSession(ctx context.Context, sql string, args ...interface{}) (rs []chunk.Row, err error) {
 	err = stm.WithNewSession(func(se sessionctx.Context) error {
-		rs, err = ExecSQL(ctx, se, sql, args...)
+		rs, err = sqlexec.ExecSQL(ctx, se, sql, args...)
 		return err
 	})
 
@@ -242,7 +227,7 @@ func (*TaskManager) CreateTaskWithSession(ctx context.Context, se sessionctx.Con
 		// disttask framework, the checked cpu-count might not right.
 		return 0, errors.Errorf("task concurrency(%d) larger than cpu count(%d)", concurrency, cpuCount)
 	}
-	_, err = ExecSQL(ctx, se, `
+	_, err = sqlexec.ExecSQL(ctx, se, `
 			insert into mysql.tidb_global_task(`+InsertTaskColumns+`)
 			values (%?, %?, %?, %?, %?, %?, %?, CURRENT_TIMESTAMP())`,
 		key, tp, proto.TaskStatePending, proto.NormalPriority, concurrency, proto.StepInit, meta)
@@ -250,7 +235,7 @@ func (*TaskManager) CreateTaskWithSession(ctx context.Context, se sessionctx.Con
 		return 0, err
 	}
 
-	rs, err := ExecSQL(ctx, se, "select @@last_insert_id")
+	rs, err := sqlexec.ExecSQL(ctx, se, "select @@last_insert_id")
 	if err != nil {
 		return 0, err
 	}
@@ -264,7 +249,7 @@ func (*TaskManager) CreateTaskWithSession(ctx context.Context, se sessionctx.Con
 // SucceedTask implements dispatcher.TaskManager interface.
 func (stm *TaskManager) SucceedTask(ctx context.Context, taskID int64) error {
 	return stm.WithNewSession(func(se sessionctx.Context) error {
-		_, err := ExecSQL(ctx, se, `
+		_, err := sqlexec.ExecSQL(ctx, se, `
 			update mysql.tidb_global_task
 			set state = %?,
 			    step = %?,
@@ -787,7 +772,7 @@ func (stm *TaskManager) UpdateSubtasksExecIDs(ctx context.Context, taskID int64,
 	}
 	err := stm.WithNewTxn(ctx, func(se sessionctx.Context) error {
 		for _, subtask := range subtasks {
-			_, err := ExecSQL(ctx, se,
+			_, err := sqlexec.ExecSQL(ctx, se,
 				"update mysql.tidb_background_subtask set exec_id = %? where id = %? and state = %? and task_key = %?",
 				subtask.ExecID,
 				subtask.ID,
@@ -819,7 +804,7 @@ func (stm *TaskManager) DeleteDeadNodes(ctx context.Context, nodes []string) err
 
 		deleteSQL.WriteString(strings.Join(deleteElems, ", "))
 		deleteSQL.WriteString(")")
-		_, err := ExecSQL(ctx, se, deleteSQL.String())
+		_, err := sqlexec.ExecSQL(ctx, se, deleteSQL.String())
 		return err
 	})
 }
@@ -881,7 +866,7 @@ func (*TaskManager) updateTaskStateStep(ctx context.Context, se sessionctx.Conte
 	}
 	// TODO: during generating subtask, task meta might change, maybe move meta
 	// update to another place.
-	_, err := ExecSQL(ctx, se, `
+	_, err := sqlexec.ExecSQL(ctx, se, `
 		update mysql.tidb_global_task
 		set state = %?,
 			step = %?, `+extraUpdateStr+`
@@ -915,7 +900,7 @@ func (*TaskManager) insertSubtasks(ctx context.Context, se sessionctx.Context, s
 			proto.TaskStatePending, proto.Type2Int(subtask.Type), subtask.Concurrency, subtask.Ordinal)
 	}
 	sb.WriteString(strings.Join(markerList, ","))
-	_, err := ExecSQL(ctx, se, sb.String(), args...)
+	_, err := sqlexec.ExecSQL(ctx, se, sb.String(), args...)
 	return err
 }
 
@@ -929,7 +914,7 @@ func (stm *TaskManager) SwitchTaskStepInBatch(
 ) error {
 	return stm.WithNewSession(func(se sessionctx.Context) error {
 		// some subtasks may be inserted by other dispatchers, we can skip them.
-		rs, err := ExecSQL(ctx, se, `
+		rs, err := sqlexec.ExecSQL(ctx, se, `
 			select count(1) from mysql.tidb_background_subtask
 			where task_key = %? and step = %?`, task.ID, nextStep)
 		if err != nil {
@@ -977,7 +962,7 @@ func (*TaskManager) splitSubtasks(subtasks []*proto.Subtask) [][]*proto.Subtask 
 func (stm *TaskManager) UpdateTaskAndAddSubTasks(ctx context.Context, task *proto.Task, subtasks []*proto.Subtask, prevState proto.TaskState) (bool, error) {
 	retryable := true
 	err := stm.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		_, err := ExecSQL(ctx, se, "update mysql.tidb_global_task "+
+		_, err := sqlexec.ExecSQL(ctx, se, "update mysql.tidb_global_task "+
 			"set state = %?, dispatcher_id = %?, step = %?, concurrency = %?, meta = %?, error = %?, state_update_time = CURRENT_TIMESTAMP()"+
 			"where id = %? and state = %?",
 			task.State, task.SchedulerID, task.Step, task.Concurrency, task.Meta, serializeErr(task.Error), task.ID, prevState)
@@ -1000,7 +985,7 @@ func (stm *TaskManager) UpdateTaskAndAddSubTasks(ctx context.Context, task *prot
 			//
 			// For now, we update nothing in proto.Task in UT's OnNextSubtasksBatch, so the AffectedRows will be 0. So UT can't fully compatible
 			// with current UpdateTaskAndAddSubTasks implementation.
-			rs, err := ExecSQL(ctx, se, "select id from mysql.tidb_global_task where id = %? and state = %?", task.ID, prevState)
+			rs, err := sqlexec.ExecSQL(ctx, se, "select id from mysql.tidb_global_task where id = %? and state = %?", task.ID, prevState)
 			if err != nil {
 				return err
 			}
@@ -1037,7 +1022,7 @@ func (stm *TaskManager) UpdateTaskAndAddSubTasks(ctx context.Context, task *prot
 					return err
 				}
 			}
-			_, err := ExecSQL(ctx, se, sql.String())
+			_, err := sqlexec.ExecSQL(ctx, se, sql.String())
 			if err != nil {
 				return nil
 			}
@@ -1076,7 +1061,7 @@ func (stm *TaskManager) CancelTask(ctx context.Context, taskID int64) error {
 
 // CancelTaskByKeySession cancels task by key using input session.
 func (*TaskManager) CancelTaskByKeySession(ctx context.Context, se sessionctx.Context, taskKey string) error {
-	_, err := ExecSQL(ctx, se,
+	_, err := sqlexec.ExecSQL(ctx, se,
 		"update mysql.tidb_global_task set state=%?, state_update_time = CURRENT_TIMESTAMP() "+
 			"where task_key=%? and state in (%?, %?)",
 		proto.TaskStateCancelling, taskKey, proto.TaskStatePending, proto.TaskStateRunning)
@@ -1100,7 +1085,7 @@ func (stm *TaskManager) IsTaskCancelling(ctx context.Context, taskID int64) (boo
 func (stm *TaskManager) PauseTask(ctx context.Context, taskKey string) (bool, error) {
 	found := false
 	err := stm.WithNewSession(func(se sessionctx.Context) error {
-		_, err := ExecSQL(ctx, se,
+		_, err := sqlexec.ExecSQL(ctx, se,
 			"update mysql.tidb_global_task set state=%?, state_update_time = CURRENT_TIMESTAMP() "+
 				"where task_key = %? and state in (%?, %?)",
 			proto.TaskStatePausing, taskKey, proto.TaskStatePending, proto.TaskStateRunning,
@@ -1123,7 +1108,7 @@ func (stm *TaskManager) PauseTask(ctx context.Context, taskKey string) (bool, er
 func (stm *TaskManager) ResumeTask(ctx context.Context, taskKey string) (bool, error) {
 	found := false
 	err := stm.WithNewSession(func(se sessionctx.Context) error {
-		_, err := ExecSQL(ctx, se,
+		_, err := sqlexec.ExecSQL(ctx, se,
 			"update mysql.tidb_global_task set state=%?, state_update_time = CURRENT_TIMESTAMP() "+
 				"where task_key = %? and state = %?",
 			proto.TaskStateResuming, taskKey, proto.TaskStatePaused,
@@ -1149,7 +1134,7 @@ func (stm *TaskManager) GetSubtasksForImportInto(ctx context.Context, taskID int
 		err error
 	)
 	err = stm.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		rs, err = ExecSQL(ctx, se,
+		rs, err = sqlexec.ExecSQL(ctx, se,
 			`select `+subtaskColumns+` from mysql.tidb_background_subtask where task_key = %? and step = %?`,
 			taskID, step,
 		)
@@ -1159,7 +1144,7 @@ func (stm *TaskManager) GetSubtasksForImportInto(ctx context.Context, taskID int
 
 		// To avoid the situation that the subtasks has been `TransferSubTasks2History`
 		// when the user show import jobs, we need to check the history table.
-		rsFromHistory, err := ExecSQL(ctx, se,
+		rsFromHistory, err := sqlexec.ExecSQL(ctx, se,
 			`select `+subtaskColumns+` from mysql.tidb_background_subtask_history where task_key = %? and step = %?`,
 			taskID, step,
 		)
@@ -1187,13 +1172,13 @@ func (stm *TaskManager) GetSubtasksForImportInto(ctx context.Context, taskID int
 // TransferSubTasks2History move all the finished subTask to tidb_background_subtask_history by taskID
 func (stm *TaskManager) TransferSubTasks2History(ctx context.Context, taskID int64) error {
 	return stm.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		_, err := ExecSQL(ctx, se, `insert into mysql.tidb_background_subtask_history select * from mysql.tidb_background_subtask where task_key = %?`, taskID)
+		_, err := sqlexec.ExecSQL(ctx, se, `insert into mysql.tidb_background_subtask_history select * from mysql.tidb_background_subtask where task_key = %?`, taskID)
 		if err != nil {
 			return err
 		}
 
 		// delete taskID subtask
-		_, err = ExecSQL(ctx, se, "delete from mysql.tidb_background_subtask where task_key = %?", taskID)
+		_, err = sqlexec.ExecSQL(ctx, se, "delete from mysql.tidb_background_subtask where task_key = %?", taskID)
 		return err
 	})
 }
@@ -1225,7 +1210,7 @@ func (stm *TaskManager) TransferTasks2History(ctx context.Context, tasks []*prot
 	return stm.WithNewTxn(ctx, func(se sessionctx.Context) error {
 		// sensitive data in meta might be redacted, need update first.
 		for _, t := range tasks {
-			_, err := ExecSQL(ctx, se, `
+			_, err := sqlexec.ExecSQL(ctx, se, `
 				update mysql.tidb_global_task
 				set meta= %?, state_update_time = CURRENT_TIMESTAMP()
 				where id = %?`, t.Meta, t.ID)
@@ -1233,7 +1218,7 @@ func (stm *TaskManager) TransferTasks2History(ctx context.Context, tasks []*prot
 				return err
 			}
 		}
-		_, err := ExecSQL(ctx, se, `
+		_, err := sqlexec.ExecSQL(ctx, se, `
 			insert into mysql.tidb_global_task_history
 			select * from mysql.tidb_global_task
 			where id in(`+strings.Join(taskIDStrs, `, `)+`)`)
@@ -1241,7 +1226,7 @@ func (stm *TaskManager) TransferTasks2History(ctx context.Context, tasks []*prot
 			return err
 		}
 
-		_, err = ExecSQL(ctx, se, `
+		_, err = sqlexec.ExecSQL(ctx, se, `
 			delete from mysql.tidb_global_task
 			where id in(`+strings.Join(taskIDStrs, `, `)+`)`)
 		return err
