+++ pkg/disttask/importinto/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "//br/pkg/lightning/backend",
         "//br/pkg/lightning/backend/external",
         "//br/pkg/lightning/backend/kv",
-        "//br/pkg/lightning/backend/local",
         "//br/pkg/lightning/checkpoints",
         "//br/pkg/lightning/common",
         "//br/pkg/lightning/config",
@@ -30,7 +29,6 @@ go_library(
         "//br/pkg/lightning/verification",
         "//br/pkg/storage",
         "//br/pkg/utils",
-        "//pkg/config",
         "//pkg/disttask/framework/handle",
         "//pkg/disttask/framework/planner",
         "//pkg/disttask/framework/proto",
@@ -42,7 +40,6 @@ go_library(
         "//pkg/domain/infosync",
         "//pkg/errno",
         "//pkg/executor/importer",
-        "//pkg/keyspace",
         "//pkg/kv",
         "//pkg/meta/autoid",
         "//pkg/metrics",
@@ -52,7 +49,6 @@ go_library(
         "//pkg/resourcemanager/pool/workerpool",
         "//pkg/resourcemanager/util",
         "//pkg/sessionctx",
-        "//pkg/sessionctx/variable",
         "//pkg/table/tables",
         "//pkg/util",
         "//pkg/util/backoff",
@@ -60,7 +56,6 @@ go_library(
         "//pkg/util/disttask",
         "//pkg/util/etcd",
         "//pkg/util/logutil",
-        "//pkg/util/mathutil",
         "//pkg/util/promutil",
         "//pkg/util/size",
         "//pkg/util/sqlexec",
@@ -71,7 +66,6 @@ go_library(
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_prometheus_client_golang//prometheus",
         "@com_github_tikv_client_go_v2//util",
-        "@io_etcd_go_etcd_client_v3//:client",
         "@org_uber_go_atomic//:atomic",
         "@org_uber_go_zap//:zap",
         "@org_uber_go_zap//zapcore",
@@ -96,14 +90,13 @@ go_test(
     embed = [":importinto"],
     flaky = True,
     race = "on",
-    shard_count = 17,
+    shard_count = 16,
     deps = [
         "//br/pkg/lightning/backend",
         "//br/pkg/lightning/backend/external",
         "//br/pkg/lightning/checkpoints",
         "//br/pkg/lightning/config",
         "//br/pkg/lightning/mydump",
-        "//br/pkg/lightning/verification",
         "//pkg/ddl",
         "//pkg/disttask/framework/planner",
         "//pkg/disttask/framework/proto",
@@ -121,7 +114,6 @@ go_test(
         "//pkg/parser/model",
         "//pkg/session",
         "//pkg/testkit",
-        "//pkg/util/logutil",
         "//pkg/util/mock",
         "//pkg/util/sqlexec",
         "@com_github_docker_go_units//:go-units",
