+++ pkg/disttask/importinto/subtask_executor_test.go
@@ -13,60 +13,3 @@
 // limitations under the License.
 
 package importinto_test
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/ngaut/pools"
-	"github.com/pingcap/failpoint"
-	verify "github.com/pingcap/tidb/br/pkg/lightning/verification"
-	"github.com/pingcap/tidb/pkg/disttask/framework/storage"
-	"github.com/pingcap/tidb/pkg/disttask/importinto"
-	"github.com/pingcap/tidb/pkg/executor/importer"
-	"github.com/pingcap/tidb/pkg/parser/model"
-	"github.com/pingcap/tidb/pkg/testkit"
-	"github.com/pingcap/tidb/pkg/util/logutil"
-	"github.com/stretchr/testify/require"
-)
-
-func TestChecksumTable(t *testing.T) {
-	ctx := context.Background()
-	store := testkit.CreateMockStore(t)
-	gtk := testkit.NewTestKit(t, store)
-	pool := pools.NewResourcePool(func() (pools.Resource, error) {
-		return gtk.Session(), nil
-	}, 1, 1, time.Second)
-	defer pool.Close()
-	mgr := storage.NewTaskManager(pool)
-
-	taskMeta := &importinto.TaskMeta{
-		Plan: importer.Plan{
-			DBName: "db",
-			TableInfo: &model.TableInfo{
-				Name: model.NewCIStr("tb"),
-			},
-		},
-	}
-	// fake result
-	localChecksum := verify.MakeKVChecksum(1, 1, 1)
-	gtk.MustExec("create database db")
-	gtk.MustExec("create table db.tb(id int)")
-	gtk.MustExec("insert into db.tb values(1)")
-	remoteChecksum, err := importinto.TestChecksumTable(ctx, mgr, taskMeta, logutil.BgLogger())
-	require.NoError(t, err)
-	require.True(t, remoteChecksum.IsEqual(&localChecksum))
-	// again
-	remoteChecksum, err = importinto.TestChecksumTable(ctx, mgr, taskMeta, logutil.BgLogger())
-	require.NoError(t, err)
-	require.True(t, remoteChecksum.IsEqual(&localChecksum))
-
-	_ = failpoint.Enable("github.com/pingcap/tidb/pkg/disttask/importinto/errWhenChecksum", `return(true)`)
-	defer func() {
-		_ = failpoint.Disable("github.com/pingcap/tidb/pkg/disttask/importinto/errWhenChecksum")
-	}()
-	remoteChecksum, err = importinto.TestChecksumTable(ctx, mgr, taskMeta, logutil.BgLogger())
-	require.NoError(t, err)
-	require.True(t, remoteChecksum.IsEqual(&localChecksum))
-}
