+++ pkg/executor/importer/BUILD.bazel
@@ -33,6 +33,7 @@ go_library(
         "//pkg/config",
         "//pkg/ddl/util",
         "//pkg/expression",
+        "//pkg/keyspace",
         "//pkg/kv",
         "//pkg/meta/autoid",
         "//pkg/parser",
@@ -58,6 +59,7 @@ go_library(
         "//pkg/util/filter",
         "//pkg/util/intest",
         "//pkg/util/logutil",
+        "//pkg/util/mathutil",
         "//pkg/util/sqlexec",
         "//pkg/util/stringutil",
         "//pkg/util/syncutil",
@@ -70,6 +72,7 @@ go_library(
         "@com_github_tikv_client_go_v2//tikv",
         "@com_github_tikv_client_go_v2//util",
         "@com_github_tikv_pd_client//:client",
+        "@io_etcd_go_etcd_client_v3//:client",
         "@org_golang_x_sync//errgroup",
         "@org_uber_go_multierr//:multierr",
         "@org_uber_go_zap//:zap",
@@ -85,18 +88,21 @@ go_test(
         "job_test.go",
         "precheck_test.go",
         "table_import_test.go",
+        "table_import_testkit_test.go",
     ],
     embed = [":importer"],
     flaky = True,
     race = "on",
-    shard_count = 20,
+    shard_count = 21,
     deps = [
         "//br/pkg/errors",
         "//br/pkg/lightning/backend/encode",
+        "//br/pkg/lightning/backend/local",
         "//br/pkg/lightning/checkpoints",
         "//br/pkg/lightning/config",
         "//br/pkg/lightning/log",
         "//br/pkg/lightning/mydump",
+        "//br/pkg/lightning/verification",
         "//br/pkg/mock",
         "//br/pkg/streamhelper",
         "//br/pkg/utils",
@@ -109,6 +115,7 @@ go_test(
         "//pkg/parser/model",
         "//pkg/planner/core",
         "//pkg/session",
+        "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/testkit",
         "//pkg/types",
@@ -120,6 +127,7 @@ go_test(
         "//pkg/util/syncutil",
         "@com_github_johannesboyne_gofakes3//:gofakes3",
         "@com_github_johannesboyne_gofakes3//backend/s3mem",
+        "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_log//:log",
