+++ pkg/util/sqlexec/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     deps = [
         "//pkg/parser",
         "//pkg/parser/ast",
+        "//pkg/parser/terror",
         "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/types",
